requirements.txt
requests==2.31.0
urllib3==1.26.16
python-whois==0.8.0
colorama==0.4.6
tqdm==4.65.0